+++ passwd/passwd.go
@@ -26,10 +26,14 @@ import (
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
+type Strength int
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
 const (
-	STRENGTH_WEAK   = iota // Only lowercase English alphabet characters
-	STRENGTH_MEDIUM        // Lowercase and uppercase English alphabet characters, digits
-	STRENGTH_STRONG        // Lowercase and uppercase English alphabet characters, digits, special symbols
+	STRENGTH_WEAK   Strength = iota // Only lowercase English alphabet characters
+	STRENGTH_MEDIUM                 // Lowercase and uppercase English alphabet characters, digits
+	STRENGTH_STRONG                 // Lowercase and uppercase English alphabet characters, digits, special symbols
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -42,12 +46,23 @@ const (
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// Encrypt creates hash and encrypts password with salt and pepper
+// Encrypt creates hash and encrypts it with salt and pepper
+// Depricated: Use Hash method instead
 func Encrypt(password, pepper string) (string, error) {
+	return Hash(password, pepper)
+}
+
+// Hash creates hash and encrypts it with salt and pepper
+func Hash(password, pepper string) (string, error) {
+	return HashBytes([]byte(password), []byte(pepper))
+}
+
+// HashBytes creates hash and encrypts it with salt and pepper
+func HashBytes(password, pepper []byte) (string, error) {
 	switch {
-	case password == "":
+	case len(password) == 0:
 		return "", errors.New("Password can't be empty")
-	case pepper == "":
+	case len(pepper) == 0:
 		return "", errors.New("Pepper can't be empty")
 	}
 
@@ -56,7 +71,7 @@ func Encrypt(password, pepper string) (string, error) {
 	}
 
 	hasher := sha512.New()
-	hasher.Write([]byte(password))
+	hasher.Write(password)
 
 	hp, err := bcrypt.GenerateFromPassword(hasher.Sum(nil), 10)
 
@@ -64,7 +79,7 @@ func Encrypt(password, pepper string) (string, error) {
 		return "", err
 	}
 
-	block, _ := aes.NewCipher([]byte(pepper))
+	block, _ := aes.NewCipher(pepper)
 	hpd := padData(hp)
 
 	ct := make([]byte, aes.BlockSize+len(hpd))
@@ -82,13 +97,18 @@ func Encrypt(password, pepper string) (string, error) {
 	return removeBase64Padding(base64.URLEncoding.EncodeToString(ct)), nil
 }
 
-// Check compares password and encrypted hash
+// Check compares password with encrypted hash
 func Check(password, pepper, hash string) bool {
-	if password == "" || hash == "" || !isValidPepper(pepper) {
+	return CheckBytes([]byte(password), []byte(pepper), hash)
+}
+
+// CheckBytes compares password with encrypted hash
+func CheckBytes(password, pepper []byte, hash string) bool {
+	if len(password) == 0 || len(hash) == 0 || !isValidPepper(pepper) {
 		return false
 	}
 
-	block, _ := aes.NewCipher([]byte(pepper))
+	block, _ := aes.NewCipher(pepper)
 	hpd, err := base64.URLEncoding.DecodeString(addBase64Padding(hash))
 
 	if err != nil {
@@ -118,34 +138,44 @@ func Check(password, pepper, hash string) bool {
 	}
 
 	hasher := sha512.New()
-	hasher.Write([]byte(password))
+	hasher.Write(password)
 
 	return bcrypt.CompareHashAndPassword(h, hasher.Sum(nil)) == nil
 }
 
 // GenPassword generates random password
-func GenPassword(length, strength int) string {
-	return getRandomPassword(length, between(strength, 0, 2))
+func GenPassword(length int, strength Strength) string {
+	return string(GenPasswordBytes(length, strength))
+}
+
+// GenPassword generates random password
+func GenPasswordBytes(length int, strength Strength) []byte {
+	return getRandomPasswordBytes(length, getStrength(strength))
 }
 
 // GetPasswordStrength returns password strength
-func GetPasswordStrength(password string) int {
-	if password == "" {
+func GetPasswordStrength(password string) Strength {
+	return GetPasswordBytesStrength([]byte(password))
+}
+
+// GetPasswordBytesStrength returns password strength
+func GetPasswordBytesStrength(password []byte) Strength {
+	if len(password) == 0 {
 		return STRENGTH_WEAK
 	}
 
 	var conditions int
 
-	if strings.ContainsAny(password, "abcdefghijklmnopqrstuvwxyz") &&
-		strings.ContainsAny(password, "ABCDEFGHIJKLMNOPQRSTUVWXYZ") {
+	if bytes.ContainsAny(password, "abcdefghijklmnopqrstuvwxyz") &&
+		bytes.ContainsAny(password, "ABCDEFGHIJKLMNOPQRSTUVWXYZ") {
 		conditions++
 	}
 
-	if strings.ContainsAny(password, "1234567890") {
+	if bytes.ContainsAny(password, "1234567890") {
 		conditions++
 	}
 
-	if strings.ContainsAny(password, _SYMBOLS_STRONG) {
+	if bytes.ContainsAny(password, _SYMBOLS_STRONG) {
 		conditions++
 	}
 
@@ -158,26 +188,61 @@ func GetPasswordStrength(password string) int {
 	switch conditions {
 	case 4:
 		return STRENGTH_STRONG
-
 	case 3:
 		return STRENGTH_MEDIUM
-
-	default:
-		return STRENGTH_WEAK
 	}
+
+	return STRENGTH_WEAK
+}
+
+// GenPasswordVariations generates password variants with possible
+// typos fixes (case swap for all letters, first leter swap, password
+// without last symbol)
+func GenPasswordVariations(password string) []string {
+	if len(password) < 6 {
+		return nil
+	}
+
+	var result []string
+
+	passwordBytes := []byte(password)
+
+	result = append(result, string(genVariantFlipAll(passwordBytes)))
+	result = append(result, string(genVariantFlipFirst(passwordBytes)))
+	result = append(result, string(genVariantTrimLast(passwordBytes)))
+
+	return result
+}
+
+// GenPasswordBytesVariations generates password variants with possible
+// typos fixes (case swap for all letters, first leter swap, password
+// without last symbol)
+func GenPasswordBytesVariations(password []byte) [][]byte {
+	if len(password) < 6 {
+		return nil
+	}
+
+	var result [][]byte
+
+	result = append(result, genVariantFlipAll(password))
+	result = append(result, genVariantFlipFirst(password))
+	result = append(result, genVariantTrimLast(password))
+
+	return result
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-func between(val, min, max int) int {
-	switch {
-	case val < min:
-		return min
-	case val > max:
-		return max
-	default:
-		return val
+func getStrength(s Strength) Strength {
+	if s < STRENGTH_WEAK {
+		return STRENGTH_WEAK
 	}
+
+	if s > STRENGTH_STRONG {
+		return STRENGTH_STRONG
+	}
+
+	return s
 }
 
 func padData(src []byte) []byte {
@@ -212,9 +277,9 @@ func removeBase64Padding(src string) string {
 	return strings.TrimRight(src, "=")
 }
 
-func getRandomPassword(length, strength int) string {
+func getRandomPasswordBytes(length int, strength Strength) []byte {
 	if length == 0 {
-		return ""
+		return nil
 	}
 
 	if strength == STRENGTH_STRONG && length < 6 {
@@ -230,27 +295,62 @@ func getRandomPassword(length, strength int) string {
 		symbols += _SYMBOLS_MEDIUM + _SYMBOLS_STRONG
 	}
 
-	for {
-		ls := len(symbols)
-		r := make([]byte, length)
+	ls := len(symbols)
+	buf := make([]byte, length)
 
+	for {
 		rand.Seed(time.Now().UTC().UnixNano())
 
 		for i := 0; i < length; i++ {
-			r[i] = symbols[rand.Intn(ls)]
+			buf[i] = symbols[rand.Intn(ls)]
 		}
 
-		if GetPasswordStrength(string(r)) == strength {
-			return string(r)
+		if GetPasswordBytesStrength(buf) == strength {
+			return buf
 		}
 	}
 }
 
-func isValidPepper(pepper string) bool {
+func isValidPepper(pepper []byte) bool {
 	switch len(pepper) {
 	case 16, 24, 32:
 		return true
 	}
 
 	return false
+}
+
+func genVariantFlipAll(password []byte) []byte {
+	result := make([]byte, len(password))
+
+	for i := 0; i < len(password); i++ {
+		result[i] = flipCase(password[i])
+	}
+
+	return result
+}
+
+func genVariantFlipFirst(password []byte) []byte {
+	result := make([]byte, len(password))
+
+	copy(result, password)
+
+	result[0] = flipCase(password[0])
+
+	return result
+}
+
+func genVariantTrimLast(password []byte) []byte {
+	return append(password[:0:0], password[:len(password)-1]...)
+}
+
+func flipCase(b byte) byte {
+	s := string(b)
+	sc := strings.ToLower(s)
+
+	if s != sc {
+		return byte(sc[0])
+	}
+
+	return byte(strings.ToUpper(s)[0])
 }